move unit tests to a separate module.
