+++ voting-contract/src/unit_tests/mod.rs
@@ -1,0 +1,8 @@
+mod bond;
+mod governance;
+mod pause;
+mod proposal;
+mod queue;
+mod reviewer;
+pub(crate) mod test_utils;
+mod votes;
